requirements.txt
fastapi
uvicorn
ffmpeg-python
pillow
pytesseract
tesseract